move lti endpoint to public
